requirements.txt
PyQt5>=5.15.0
openai-whisper>=20231117
torch>=2.0.0
numpy>=1.20.0
requests>=2.25.0
pillow>=8.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0 